feat(hypo): fix int/long ABI mismatch of hypo_ flag arguments
hypoplas.cc passed int find_local_sv[1]/options_nonlocal[1] (4 B) to
hypo_, whose f2c-style pure-C port declares them long int * (8 B on
LP64): hypo.c:165 and sigma_ (hypo.c:620-621) load 8 B from 4-B stack
objects -> deterministic stack-buffer-overflow exposed by ASan (make
asan aborted hypo1-4). Declare both flags as long int in hypoplas.cc
(extern prototype and locals), consistent with the remaining INTEGER
arguments of the port (nhis, ndata, use_pres, use_epi, hypo_type) and
with the typedef long int integer convention of tochnog.h.

Verified: make asan suite 16/16 (was 12/16); corpus hypo1 and suite
hypo1 run under tochnog-asan with no AddressSanitizer report; default
build numeric output byte-identical before/after on hypo1-4/7-9/12
(the OOB garbage read was zero padding, so it did not fire spurious
branches in the default -O1 build).

+++ hypoplas.cc
@@ -34,7 +34,7 @@ extern "C"
     double *old_epi, double *new_epi,
     long int *use_pres, long int *use_epi, long int *hypo_type,
     double *softvar_nonloc, double *softvar_loc,
-    int *find_local_sv, int *options_nonlocal );
+    long int *find_local_sv, long int *options_nonlocal );
 
 extern "C" 
   void masin_umat( double *stress, double *statev, double *ddsdde,
@@ -71,7 +71,7 @@ void hypoplasticity( long int element, long int gr,
     length_lowangles=0, length_wolfersdorff=0, length_intergranularstrain=0, 
     pressure_dependent_void_ratio=0, idum[1],
     ndata[1], nhis[1], use_epi[1], use_pres[1], hypo_type[1];
-  int find_local_sv[1], options_nonlocal[1];
+  long int find_local_sv[1], options_nonlocal[1];
   double ddum[1], cohesion[1], epi_R[1], epi_mr[1], 
     epi_mt[1], epi_betar[1], epi_chi[1], time[1], dtime[1],
     softvar_nonloc[1], softvar_loc[1],
